Crate sprintf

Libc s(n)printf clone written in Rust, so you can use printf-style formatting without a libc (e.g. in WebAssembly).

Note: You’re probably better off using standard Rust string formatting instead of thie crate unless you specificaly need printf compatibility.

It follows the standard C semantics, except:

Locale-aware UNIX extensions (' and GNU’s I) are not supported.

%a/%A (hexadecimal floating point) are currently not implemented.

Length modifiers (h, l, etc.) are checked, but ignored. The passed type is used instead.

Usage example:

use sprintf::sprintf;
let s = sprintf!("%d + %d = %d\n", 3, 9, 3+9).unwrap();
assert_eq!(s, "3 + 9 = 12\n");

The types of the arguments are checked at runtime.

Macros

sprintf

Format a string. (Roughly equivalent to snprintf or asprintf in C)

Structs

ConversionSpecifier

Parsed printf conversion specifier

Enums

ConversionType

Printf data type

NumericParam

Width / precision parameter

PrintfError

Error type

Traits

Printf

Trait for types that can be formatted using printf strings

Functions

vsprintf

Format a string. (Roughly equivalent to vsnprintf or vasprintf in C)

Type Aliases

Result